+++ src/it/pollutionProtectionInstrument/src/main/java/A.java
@@ -1,0 +1,5 @@
+public class A {
+    public boolean a() {
+        return true;
+    }
+}